Created seperate component for node tree details
